apps/api/pyproject.toml
[project]
name = "ri-assessments-api"
version = "0.1.0"
description = "Revenue Institute Assessments Platform — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "supabase>=2.9.0",
  "python-jose[cryptography]>=3.3.0",
  "anthropic>=0.40.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.12",
  "redis>=5.1.0",
  "psycopg[binary]>=3.2.0",
]

[dependency-groups]
dev = [
  "ruff>=0.7.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "datamodel-code-generator>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ri_assessments_api"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
